# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_logic3579_optimus_internal_infra_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      message_key:
        type: string
    type: object
...
  internal_modules_observability_datasource.CreateRequest:
    properties:
      auth_type:
        enum:
        - none
        - basic
        - bearer
        type: string
      base_url:
        type: string
      cluster_id:
        type: integer
      custom_ca_pem:
        type: string
      description:
        maxLength: 4096
        type: string
      http_credential_id:
        type: integer
      name:
        maxLength: 128
        type: string
      tls_skip_verify:
        type: boolean
    required:
    - auth_type
    - base_url
    - name
    type: object
  internal_modules_observability_datasource.Detail:
    properties:
      auth_type:
        type: string
      base_url:
        type: string
      cluster:
        $ref: '#/definitions/internal_modules_observability_datasource.NamedRef'
      created_at:
        type: string
      created_by_user_id:
        type: integer
      description:
        type: string
      has_custom_ca:
        type: boolean
      http_credential:
        $ref: '#/definitions/internal_modules_observability_datasource.NamedRef'
      id:
        type: integer
      name:
        type: string
      tls_skip_verify:
        type: boolean
      updated_at:
        type: string
    type: object
  internal_modules_observability_datasource.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_modules_observability_datasource.Detail'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_modules_observability_datasource.NamedRef:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  internal_modules_observability_datasource.TestResponse:
    properties:
      reachable:
        type: boolean
      version:
        type: string
    type: object
  internal_modules_observability_datasource.UpdateRequest:
    properties:
      auth_type:
        enum:
        - none
        - basic
        - bearer
        type: string
      base_url:
        type: string
      clear_cluster:
        type: boolean
      clear_custom_ca:
        type: boolean
      clear_http_credential:
        type: boolean
      cluster_id:
        type: integer
      custom_ca_pem:
        type: string
      description:
        maxLength: 4096
        type: string
      http_credential_id:
        type: integer
      name:
        maxLength: 128
        type: string
      tls_skip_verify:
        type: boolean
    type: object
...
  internal_modules_rbac.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 128
        minLength: 8
        type: string
      old_password:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - new_password
    - old_password
    type: object
  internal_modules_rbac.UpdateMeRequest:
    properties:
      avatar_url:
        maxLength: 512
        type: string
      display_name:
        maxLength: 128
        type: string
      email:
        maxLength: 128
        type: string
    type: object
...
  internal_modules_role.SetPermissionsRequest:
    properties:
      permission_codes:
        items:
          type: string
        type: array
    required:
    - permission_codes
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    P0 admin backend for Optimus — auth, RBAC, users, roles, permissions, menus, audit.
    All authenticated endpoints expect `Authorization: Bearer <access_token>`.
  title: Optimus Admin API
  version: "1.0"
paths:
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user profile
      tags:
      - me
    put:
      consumes:
      - application/json
      parameters:
      - description: profile patch
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - me
  /me/menus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user menu tree
      tags:
      - me
  /me/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: old/new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Change current user password
      tags:
      - me
  /me/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user permission codes
      tags:
      - me
...
  /observability/datasources:
    get:
      parameters:
      - description: name search
        in: query
        name: q
        type: string
      - description: authentication type
        in: query
        name: auth_type
        type: string
      - description: cluster ID
        in: query
        name: cluster_id
        type: integer
      - default: 1
        description: page
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_datasource.ListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List observability data sources
      tags:
      - observability
    post:
      consumes:
      - application/json
      parameters:
      - description: data source
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_observability_datasource.CreateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_datasource.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Create observability data source
      tags:
      - observability
  /observability/datasources/{id}:
    delete:
      parameters:
      - description: data source ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete observability data source
      tags:
      - observability
    get:
      parameters:
      - description: data source ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_datasource.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Get observability data source
      tags:
      - observability
    put:
      consumes:
      - application/json
      parameters:
      - description: data source ID
        in: path
        name: id
        required: true
        type: integer
      - description: data source
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_observability_datasource.UpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_datasource.Detail'
              type: object
      security:
      - BearerAuth: []
      summary: Update observability data source
      tags:
      - observability
...
  /observability/datasources/{id}/test:
    post:
      parameters:
      - description: data source ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_observability_datasource.TestResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Test observability data source
      tags:
      - observability
...
  /permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - permissions
...
  /roles/{id}/permissions:
    put:
      consumes:
      - application/json
      parameters:
      - description: role ID
        in: path
        name: id
        required: true
        type: integer
      - description: permission codes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_role.SetPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Set role permissions
      tags:
      - roles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"